rohanpm: order of the libs passed to ld matters
rohanpm: "The linker will search an archive only once, at the location where it is specified on the command line"

rohanpm: I think that means if libfooer uses libcommon, then -lcommon must come after -lfooer on the command-line

rohanpm: (guessing) linker is trying to link toberun. It processes -lfooer on the command-line, finds & opens libfooer.a and tries to resolve any unresolved symbols used in there. It couldn't resolve foo::someFooFunc() which comes from libcommon.a ; and the libfooer.a retained enough information to tell exactly what line of the .cpp was referring to it, so that was included in the message